Removing now redundant cast.

+++ src/main/java/org/graylog2/alarmcallbacks/jabber/JabberAlarmCallback.java
@@ -60,7 +60,7 @@ public class JabberAlarmCallback implements AlarmCallback {
 
         final ConnectionConfiguration connectionConfiguration = new ConnectionConfiguration(
                 config.getString(CK_HOSTNAME),
-                (int) config.getInt(CK_PORT),
+                config.getInt(CK_PORT),
                 serviceName
         );
